requirements.txt
click==7.1.2
Django==2.2.12
djangorestframework==3.12.2
fuzzywuzzy==0.18.0
joblib==1.0.0
nltk==3.5
python-Levenshtein==0.12.0
pytz==2020.5
regex==2020.11.13
sqlparse==0.4.1
tqdm==4.55.1